stock/requirements.txt
lark-oapi>=1.0.0
playwright>=1.40.0
pyyaml>=6.0
